Support generic listener in config file
This allows creation of multiple listeners, with different protocols
and/or SSL certificates.

Related change: settings is now a class that is copyable and assignable,
and is done so to each thread on reload.

Semi-related fix: fix crash in quit when multiple threads initiated it.
This came to light when testing the auth plugin settings.