Lock (QoS data of) sessions during copying/saving
The only mutable session data of a client is QoS related, so when we're
copying sessions (for saving them), we need to lock the QoS data,
because that gets modified from active client traffic in worker threads.

Note: not super well tested at this point, nor was I ever able to
trigger actual errors despite long stress testing, so it's a theoretical
fix.