Fix invalid API in rwlockguard
The code allowed one to try to upgrade a read lock to a write lock,
which would/could fail.

In the existing code, the only attempted double lock was a write lock,
so this change doesn't actually fix anything, but it's clearer, and we
can make do with the lock that was removed. It just takes a bit more
overhead on app start-up, to place a lock per message.