Allow MQTT5 non-clean start empty client ID
When you have the server generate a client ID: "It MUST then process the
CONNECT packet as if the Client had provided that unique ClientID, and
MUST return the Assigned Client Identifier in the CONNACK packet".

So, this means allowing an empty client ID, even for non-clean start,
should be allowed.