Add option to expire retained messages
This also includes a tree-purge that needed to happen, whether you
expire messages based on time or not (because nodes can get orphaned).

The tasks is performed in one of the worker threads.

This also fixes a bug: retained messages were erased on-demand before,
under a read lock, which is obviously wrong.

The purge action is given a time budget, with appropriate log warning.
While the task is running, one thread holds a global lock, so we want to
avoid doing this too long.

This also fixes a logic in age calculation, affecting the stored
sessions.