Reduce write locking on setting retained messages
Assuming that the payload of a topic changes many times, it makes sense
to only place a write lock when the tree node needs changing i.e. when
setting a message for the first time. This means most of the work can be
done under a global read lock.

The setting of the actual payload is still done under mutex; one per
tree node.

This will allow a much greater parallism on setting retained messages.