Add `--fail-on-doc-failure` flag to `build.sh` script
@halfgaar came up with this flag in private conversation, and I thought
it would be a good idea to implement it to have a mechanism to ensure
that release pipelines always sport up-to-date man page builds.

Another minor change is that I substituted the `.PRECIOUS` commentd-out
pseudo-target in the `Makefile` with `.SECONDARY`.  The latter is the
“correct” way to keep intermediate files from being deleted, without
getting attached to _faulty_, incomplete intermediate files resulting
from crashes.