Access to payload in plugin
Used a string_view, because an alterable payload is just a bridge too far with
regards to certain optimizations that are in place. Having a read-only version
is better than nothing, so I decided this is the way to go.

I also removed the CI build for Ubuntu 18.04. It's about to be EOL and
the runner will be removed. This allows the use of c++17.

https://github.com/halfgaar/FlashMQ/issues/33