Fix use-after-free in fuzz mode
Because packets have a shared pointer to a client, we need to make sure
packets are destroyed right after using, otherwise the client destructor
gets called not on destruction of client, but of the packet. And in this
case, that happened after the authentication object was already
destroyed.

This only impacted fuzzing mode, and was detecting using address
sanitizing.