Add support for bridge mode
A bridge is essentially a normal client, but then FlashMQ takes care of
sending the connect packet and using a timer to keep it connected.

The defacto bridge protocol version bit in the connect packet is
supported. It takes care of MQTT3 'no local' and 'retained as
published'. MQTT5 can do that natively.