Fix not reading websocket buffer on certain frame sizes
The while condition that caused the read to be skipped was previously
added as a fix for spinning when the websocket length header was
corrupt. This was found with fuzzing at the time. However, new fuzzing
attempts fail to reproduce that.

The while condition didn't make sense, because it's very possible to
have frames bigger than the current buffer.