Fix crash on corrupt MQTT packets
When clients send certain corrupt MQTT packets, this was detected very
quickly and caused the thread to close and abandon the client while the
main loop was still busy registering it. This race condition could cause
the client destructor to be called in the main thread, instead of the
worker thread. That in turn, dereferenced nullptrs, trying to obtain
thread global objects.

It's fixed by MOVING the client to the thread, instead of copying.

Detected by Benedikt Brandmaier using rusty-FUME.