Fix losing QoS packets on double reconnect
It removed QoS packets from the queue as it resent them on reconnect. We
need to keep them in the queue and remove them with the ack as normal.

This fixes QoS resending for clients with very bad connections, where
the connection immediately goes bad again after a reconnection.

The QoS 2 PUBREL packets didn't have this problem.