Fix deleting session with the same ID on reconnect
Reconnecting with the same client ID that still exists on the server
results disconnection and deletion of the old client, and it erroneously
queued a deletion of the session with that ID, resulting in deleting the
freshly created session that replaces the old one.

Fixed by clearing the session of the client that's being disconnected.

This fixes losing subscriptions in these scenarios.

Small note about session destructor logging: previously this was removed
because it caused this line to be printed on copied sessions on saving
them to disk. But, we don't copy anymore, so now we can log again.