Don't reset session of existing client on connect
This was originally done to prevent deleting the existing AND newly
created session on connecting with the same ID. See commit eebab1cb07.
But, that was subsequently also fixed by making the removeSession()
smarter, so we don't need to reset the original session.

Deleting the session of the existing client had the side effect of
handling delayed packets of clients without a session, resulting in a
null dereference. See previous commit.