Defer giving retained messages until done
This fixes long blockages when someone subscribes to '#' when there are
many retained messages, or many empty nodes in the retained tree.

Various tests have been conducted with a tree of 10M messages, over slow
lines, with and without QoS, and the result is always deterministic,
aside from the delivery order.