Add option to set listener socket(s) to `TCP_NODELAY` mode
How I came to the idea for this new option was when I learned that
`curl`, since version 7.50.2, sets `TCP_NODELAY` by default.

`TCP_NODELAY` _may_ decrease latency and may thus be worthwhile to be
able to experiment with in some scenarios, to try if its benefits don't
outweigh its main tradeoff (i.e., more, smaller TCP messages).

The reason that I added this as a boolean setting is that, as far as
I can tell, there are only three modes to choose from, and I guessed
that Wiebe is not interested in using `TCP_CORK` in the future:

1. The default is that the TCP layer uses Nagle's algorithm to collect
   what would become multiple, small TCP packets into a larger packet
   until acknowledgement has been received for the last sent message.
2. `TCP_NODELAY` sends out TCP packets immediately, possible reducing
   latency, but at the likely cost of efficiency.
3. `TCP_CORK` does such buffering explicitly, waiting for
   `setsockoption()` from the application to flush the buffer.

If I am wrong and there _is_ a potential future use for `TCP_CORK`,
this option should probably be changed into an enum.

I have done a superficial test of this new feature using `mosquitto_sub`
and `mosquitto_pub`, only to the extent to check if a listener with
`tcp_nodelay true` still functions.  What I did not do is to trace what
actually happens on the TCP level.

I do not whether and where this needs to be added to the automated test
suite.