Add `tcp_nodelay` bridge parameter to match
As per Wiebe's request, this commit adds a `tcp_nodelay` bridge
parameter to match the previously-added same-named listener parameter.

In `Client::connectToBridgeTarget()`, I took the liberty to remove the
`if (rc == 0)` condition.  It appeared redundant to me, since
`connect()` will always return a value `<= 0`.  To make/keep that clear,
I replaced the `if (rc == 0)` with an `assert(rc == 0)`.

For `BridgeConfig`, I've opted to _not_ use a getter to access the
value of its new `BridgeConfig.tcpNoDelay` member variable, to be
consistent with style of how most other member variables of this
class are used.  (This is contrary to what I did for the
`Listener.tcpNoDelay` member, where I did add a getter, because that
seemed to be how the other boolean members of that other class were used
as well.)