Fix not queueing keep-alives when the server doesn't respond
Likely the fact that queueing repeating keepalive checks was done in
CONNACK versus right from the start was to use a possibly updated
keepalive value, but that will happen anyway if we just always requeue
the check.

This fixes reconnect issues when the server accepted the TCP connection
but then silently dropped it.