Refactor the copy factory and mqtt packet construction to not alter source
This fixes a wrong implementation of an optimization.

The optimization of reusing the source (and avoiding making identical
copies for subscribers) should really only be done when dealing with a
packet, not a publish object. By allowing overrides to the MqttPacket
constructor, we can actually even remove an extraneous copy of a publish
object. So, this is cleaner and faster.

This fixes QoS interference of multiple receivers of wills. See test in
previous commit.