Don't lock session.client shared pointer for clients without access
When you have many subscribers that don't have access, the code before
the ACL check gets called at a multiplied count of message rate, so
don't do any extraneous work here.

The only reason it was needed, was for the effective retain, and I think
it was actually a bug that we used that value. The ACL check should be
with the retain flag of the message, not depending on whether the
receiving client connected with 'retain unavailable'.