Do optimistic insertion into list of subscribers
We remove it again when we don't need it.

This saves a copy of the shared pointer. During profiling, the
std::shared_ptr<Session> destructor got a fair amount of CPU, so it's
worth while to optimize that.

Before, we always called a destructor. Now only when we don't need the
session.

The only extra overhead is pop_front(), which is very small.

This also fixes the no_local option for shared subscriptions.