Fix handling packets already received after CONNECT in async auth
The spec supports already sending subsequent packets after the initial
CONNECT. When handling it async, it would discard them. This fix queues
them for processing on successful authentication.

This refactors it such that the continuation of authentication is not
done as a task, but in the event loop. This is necessary to do error
handling as a normal client. Otherwise there's just an error logged
about the queued task and the client is left in a weird state.