Mostly cosmetic `man 5 flashmq.conf` improvements
When there were alternative values for an option, these were oft
(1) separated by a slash, and (2) placed within a single `<replaceable>`
element. This (1) made them look very similar to path values, and (2)
made the values and their separator all end up with the same formatting.

Also, there were a couple of options for which the example/enum values
were not included in the `<term>` at all, but instead included in a
list, which, though not wrong in any way, I found to be inconsistent.