Fix possible hanging SSL read loop
Actual case not reproduced, but it was confirmed that doing an extra
SSL_read() on a socket that had already returned EOF made it get stuck
in a loop.

This refactor makes the code more in line with the documentation for
OpenSSL 3, instead of OpenSSL 1, meaning we don't communicate the
SSL_read() error upwards, but stick to our own convention.