In test client, hold lists of received packets in MutexOwned
We did unsafe thread access. This changes ensures we lock the mutex.

This may fix some occassional weird errors that have been observed when
running tests in a loop all day. It's better way.

This also fixes testExtendedReAuthTwoStep, which did not properly look
at reauthFinishData. It was noticed because of adding more local scope
for locks.