Add 'connection_count' bridge option for load balancing
Uses MQTT5 shared subscriptions, with a bit of custom behavior
communicated using user properties.

It allows setting up multiple connections to the same subscription path
to load-balance over. This is especially important for the side that
receives publishes, because that's where most of the load is (lookin up
subscribers and such).