Avoid cross-thread problems using bridge prefix paths
Because the std::optional won't be marked as 'engaged' until after
construction, we will never use it while being constructed. And because
we never change the contained value, we share it without locking.

It's a work-around to haing a (shared) mutex on the prefixes, which would
only add the extra feature of being able to change the prefix paths with
SIGHUP, which is not worth the overhead.