Fix flaky 'test_retained_removed()'
The assert sometimes failed with these values:

  Payload: ''. Retain: 0. Access: 1 (read)

That's because the receiver test client connects too fast and is still
able to see the publish of the empty payload as normal message.

By publishing with QoS, we actually wait for the ack before connecting
the receiving test client.

Fixes https://github.com/halfgaar/FlashMQ/issues/130