Reduce log noise and change log levels of frequent events
This refactors the logging to:

* Don't log any random client event that causes disconnect as ERROR.
* Reduce printed log lines.

In normal server operation, misbehaving clients are normal. Logging all
of those as errors makes no sense. So now, client errors are just logged
as a normal client disconnect reason, not as an extra ERROR line.

The outgoing client errors are set as WARNING, because many of those are
under our control.

The idea is that ERROR log entries should be actual errors that someone
is interested in.